requirements.txt
streamlit==1.10.0
numpy==1.22.3
Pillow==9.2.0
tensorflow==2.8.0
